Refactorizando la lase AccountViewModel enviando a la carpeta ViewModels las clases que tenia contenidas.
