+++ Code/GitHub/GitHub/Controllers/AccountController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using GitHub.Models;
+using GitHub.ViewModels;
 
 namespace GitHub.Controllers
 {
