+++ Code/GitHub/GitHub/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model GitHub.Models.ForgotPasswordViewModel
+﻿@model GitHub.ViewModels.ForgotPasswordViewModel
 @{
     ViewBag.Title = "Forgot your password?";
 }
