+++ Code/GitHub/GitHub/Views/Account/Login.cshtml
@@ -1,5 +1,5 @@
 ﻿@using GitHub.Models
-@model LoginViewModel
+@model GitHub.ViewModels.LoginViewModel
 @{
     ViewBag.Title = "Log in";
 }
