+++ Code/GitHub/GitHub/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model GitHub.Models.RegisterViewModel
+﻿@model GitHub.ViewModels.RegisterViewModel
 @{
     ViewBag.Title = "Register";
 }
