+++ Code/GitHub/GitHub/Views/Account/ResetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model GitHub.Models.ResetPasswordViewModel
+﻿@model GitHub.ViewModels.ResetPasswordViewModel
 @{
     ViewBag.Title = "Reset password";
 }
