+++ Code/GitHub/GitHub/Views/Account/VerifyCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model GitHub.Models.VerifyCodeViewModel
+﻿@model GitHub.ViewModels.VerifyCodeViewModel
 @{
     ViewBag.Title = "Verify";
 }
